mail-function: move relevant code into index.js
The code in index.js was just a poc, now we need the real implementation.
